# web-portal/backend/openapi.yaml
openapi: 3.0.0
info:
  title: VetsNPets Backend API
  description: Backend service for the VetsNPets frontend.
  version: "1.0"
servers:
  - url: http://localhost:8080
    description: Development server
paths:
  /create-appointment:
    post:
      summary: Create an appointment
      description: Forwards the appointment creation request to the Appointment service with proper authentication.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AppointmentRequest'
      responses:
        '200':
          description: Appointment created successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AppointmentResponse'
        '401':
          description: Authentication failed
        '500':
          description: Internal server error
  /appointments:
    get:
      summary: Get appointments by email
      description: Fetches appointments for the given email. Supports filtering by the 'upcoming' query parameter.
      parameters:
        - in: query
          name: email
          schema:
            type: string
            format: email
          required: true
          description: Email address to filter appointments
        - in: query
          name: upcoming
          schema:
            type: string
            enum: ['true', 'false']
          required: false
          description: Optional filter for fetching only upcoming appointments
      responses:
        '200':
          description: A list of appointments
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/AppointmentResponse'
        '400':
          description: Missing or invalid parameters
        '401':
          description: Authentication failed
        '500':
          description: Internal server error
components:
  schemas:
    AppointmentRequest:
      type: object
      properties:
        name:
          type: string
          example: John Doe
        service:
          type: string
          example: Dental Checkup
        phoneNumber:
          type: string
          example: '+1234567890'
        appointmentDate:
          type: string
          format: date-time
          example: '2023-04-30T14:00:00Z'
        email:
          type: string
          format: email
          example: john.doe@example.com
      required:
        - name
        - service
        - phoneNumber
        - appointmentDate
        - email
    AppointmentResponse:
      type: object
      properties:
        id:
          type: integer
          example: 1
        name:
          type: string
          example: John Doe
        service:
          type: string
          example: Dental Checkup
        phoneNumber:
          type: string
          example: '+1234567890'
        appointmentDate:
          type: string
          format: date-time
          example: '2023-04-30T14:00:00Z'
        email:
          type: string
          format: email
          example: john.doe@example.com
        status:
          type: string
          example: Confirmed
      required:
        - id
        - name
        - service
        - phoneNumber
        - appointmentDate
        - email
        - status